+++ lib/astro_chart.rb
@@ -28,6 +28,7 @@ require_relative "astro_chart/transit_timing"
 require_relative "astro_chart/solar_arc"
 require_relative "astro_chart/lunar_return"
 require_relative "astro_chart/horary"
+require_relative "astro_chart/planetary_hours"
 
 module AstroChart
 end